Removes config file capability
just realized config files were overkill for a script this tiny.

As unlikely as it is, if any one other than me uses this script
they're most likely going to be capable enough of editing it to
fit their needs, so there's no need to bother with a proper config
file setup. All the config stuff is handled by variables at the top of
the script. :)

Apart from that, i made the post subtitles optional per post. also
tried error handling but that doesn't work very well right now.